Changes more counters and sizes to size_t

+++ src/hotspot/share/cds/cdsHeapVerifier.hpp
@@ -40,7 +40,7 @@ class CDSHeapVerifier : public KlassClosure {
   class CheckStaticFields;
   class TraceFields;
 
-  int _archived_objs;
+  size_t _archived_objs;
   int _problems;
 
   struct StaticFieldInfo {
